pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
]
